The First Step
Following are the things that have been done/achieved in the 1st commit ( officially 3rd ) of the Arcade Warrior game.

I have created the project with the environment and the battle arena with the free source assets through the unity hub store. ( more improvement on design later on )

A player object model has also been added to the game in middle of arena.

2 scripts have been coded.

1st one is based on the player movement and the rotation.
-help player move around with wasd keys

2nd one is the camera following player's origin point at a fixed location.